argument parsing moved (main->if)
Before PR#3 commit "part_c_vpol.py and housekeeping", Ezio left a comment suggesting
to move the argument parsing stuff under the if statement (ie. if __name__=='__main__':)

This is an attempt to do just that for sspermissioncheck.py

The script works; will check with Ezio that this is correct before changing part_c_vpol.py as well.